fix docs and move filesize check out of checks/vendorspecific/googlefonts
