+++ Lib/fontbakery/checks/vendorspecific/googlefonts/metadata.py
@@ -4,7 +4,7 @@ import os
 from fontbakery.constants import NameID, RIBBI_STYLE_NAMES
 from fontbakery.prelude import check, Message, INFO, PASS, FAIL, WARN, SKIP, FATAL
 from fontbakery.utils import exit_with_install_instructions, show_inconsistencies
-from fontbakery.checks.googlefonts.glyphset import can_shape
+from fontbakery.checks.vendorspecific.googlefonts.glyphset import can_shape
 
 
 @check(
