+++ Lib/fontbakery/profiles/notofonts.py
@@ -1,6 +1,6 @@
 PROFILE = {
     "configuration_defaults": {
-        "googlefonts:file_size": {
+        "file_size": {
             "WARN_SIZE": 1 * 1024 * 1024,
             "FAIL_SIZE": 16 * 1024 * 1024,
         },
