+++ docs/source/fontbakery/checks/adobefonts.rst
@@ -2,5 +2,5 @@
 Checks to support the Adobe Fonts profile
 #########################################
 
-.. automodule:: fontbakery.checks.adobefonts
+.. automodule:: fontbakery.checks.vendorspecific.adobefonts
     :members:
