+++ docs/source/fontbakery/checks/fontbureau.rst
@@ -2,5 +2,5 @@
 Checks to support the Fontbureau profile
 ########################################
 
-.. automodule:: fontbakery.checks.fontbureau
+.. automodule:: fontbakery.checks.vendorspecific.fontbureau
     :members:
