+++ docs/source/fontbakery/checks/fontwerk.rst
@@ -2,5 +2,5 @@
 Checks to support the Fontwerk profile
 ######################################
 
-.. automodule:: fontbakery.checks.fontwerk
+.. automodule:: fontbakery.checks.vendorspecific.fontwerk
     :members:
