+++ docs/source/fontbakery/checks/googlefonts.rst
@@ -5,149 +5,149 @@ Checks to support the Google Fonts profile
 Base checks
 ***********
 
-.. automodule:: fontbakery.checks.googlefonts
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts
     :members:
 
 Axis registry checks
 ********************
 
-.. automodule:: fontbakery.checks.googlefonts.axisregistry
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.axisregistry
     :members:
 
 Color font checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.color
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.color
     :members:
 
 Copyright checks
 ****************
 
-.. automodule:: fontbakery.checks.googlefonts.copyright
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.copyright
     :members:
 
 Description checks
 ******************
 
-.. automodule:: fontbakery.checks.googlefonts.description
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.description
     :members:
 
 Family checks
 *************
 
-.. automodule:: fontbakery.checks.googlefonts.family
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.family
     :members:
 
 GDEF table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.gdef
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.gdef
     :members:
 
 glyf table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.glyf
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.glyf
     :members:
 
 Glyphset checks
 ***************
 
-.. automodule:: fontbakery.checks.googlefonts.glyphset
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.glyphset
     :members:
 
 GPOS table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.gpos
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.gpos
     :members:
 
 GSUB table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.gsub
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.gsub
     :members:
 
 head table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.head
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.head
     :members:
 
 Hinting checks
 **************
 
-.. automodule:: fontbakery.checks.googlefonts.hinting
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.hinting
     :members:
 
 Regressions checks against fonts hosted on fonts.google.com
 ***********************************************************
 
-.. automodule:: fontbakery.checks.googlefonts.hosted
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.hosted
     :members:
 
 License checks
 **************
 
-.. automodule:: fontbakery.checks.googlefonts.license
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.license
     :members:
 
 meta table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.meta
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.meta
     :members:
 
 Metadata checks
 ***************
 
-.. automodule:: fontbakery.checks.googlefonts.metadata
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.metadata
     :members:
 
 name table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.name
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.name
     :members:
 
 OS/2 table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.os2
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.os2
     :members:
 
 Repository checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.repo
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.repo
     :members:
 
 STAT table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.stat
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.stat
     :members:
 
 Subsetting checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.subsets
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.subsets
     :members:
 
 Font table checks
 *****************
 
-.. automodule:: fontbakery.checks.googlefonts.tables
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.tables
     :members:
 
 Variable font checks
 ********************
 
-.. automodule:: fontbakery.checks.googlefonts.varfont
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.varfont
     :members:
 
 Vertical metrics checks
 ***********************
 
-.. automodule:: fontbakery.checks.googlefonts.vmetrics
+.. automodule:: fontbakery.checks.vendorspecific.googlefonts.vmetrics
     :members:
