+++ docs/source/fontbakery/checks/microsoft.rst
@@ -1,0 +1,6 @@
+#######################################
+Checks to support the Microsoft profile
+#######################################
+
+.. automodule:: fontbakery.checks.vendorspecific.microsoft
+    :members:
