+++ docs/source/fontbakery/checks/notofonts.rst
@@ -2,5 +2,5 @@
 Checks to support the Noto fonts profile
 ########################################
 
-.. automodule:: fontbakery.checks.notofonts
+.. automodule:: fontbakery.checks.vendorspecific.notofonts
     :members:
