--- docs/source/fontbakery/checks/universal.rst
@@ -1,6 +1,0 @@
-#########################################################
-Checks for commonly agreed font production best practices
-#########################################################
-
-.. automodule:: fontbakery.checks.universal
-    :members:
