+++ tests/test_checks_googlefonts_conditions.py
@@ -1,7 +1,7 @@
 import pytest
 from fontTools.ttLib import TTFont
 
-from fontbakery.checks.googlefonts.glyphset import is_icon_font
+from fontbakery.checks.vendorspecific.googlefonts.glyphset import is_icon_font
 from fontbakery.codetesting import TEST_FILE, MockContext, MockFont
 from fontbakery.fonts_profile import setup_context
 from fontbakery.testable import Font
